requirements.txt
matplotlib==3.9.4
opencv-python==4.11.0.86
requests==2.32.3
numpy==2.0.2
tqdm==4.66.5
python-dotenv==1.0.1

# ML pipeline — install PyTorch separately per platform:
#   CPU (this machine):  pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#   CUDA 12.x (3080 PC): pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch
torchvision
Pillow
scikit-learn
boto3
